Changed ratio to 10% of previous ratio
After changing harvesters to gather 10% of what intakes would take of IntakeAir if ExpVapour is 10% of an atmosphere, I changed the engine ratios to a baseline of 0.375:1 to match pure ethane combustion with oxygen. Engines are still constrained more by velCurve and atmCurve, so this shouldn't break the engines unless ExpVapour is under 1.5% of an atmosphere. The turboFanSize2Ox's harvester is also adjusted to match the other harvesters.

Engines that have visible stock plumes have new plumes representing methane / ethane combustion with oxygen. Using Squad stock plume from the Vector engine for this purpose.